HOCS-2444 Add health-checks to Compose; add Readme
This commit improves the Docker Compose experience for Frontend
somewhat, by adding documentation for Docker Compose, and making sure
that Docker brings the containers up in a sensible order (postgres
before info-service, for example; or Localstack before docs). I think
this has always been the intention with the Docker Compose file, as the
`depends_on` properties were already defined -- however, they only
waited for containers to *start*, and didn't wait for the containers to
be *ready* (which is now the new behaviour).

The feature this relies on was removed from Docker 3, but we're not
using any Docker 3 features so it's safe to drop it back down to Docker
2. (We're not using Docker Swarm, you see.) Any docker-compose version
devs are already using will still support this older version of the
syntax.

We add a Readme as this is likely the first part of DECS a new developer
will touch, so this improves the onboarding slightly.

Startup times for Docker Compose might be increased by a minute or so
compared to previously, as we now wait for Localstack to come up safely
before running most of the services. This commit might also mean that
there are more errors than previously -- Compose might complain about
unhealthy containers -- but this is actually just Docker being louder:
these errors will have still existed, but you wouldn't have noticed them
previously.

This commit probably removes the requirement to have separate
"infrastructure" and "services" startup scripts, but I've left them in
so I don't disrupt existing developer workflow.