Improve environment and depends_on
A number of the environment variables that are currently passed in 
don't need to be. This is either because sensible defaults are set or 
the variables are not used. Further tickets should be created to 
identify any changing defaults required. 
Alongside this, the workflow service is not required for at least the 
initial use case of management-ui so this is removed. We do 
however require that the s3 buckets in localstack have been created 
to use the system effectively.