HOCS-3205: Add Management UI to docker compose  (#9)
* Prettifier changes

We have a range of different changes that prettifier has changed. 
This commit applies all of these changes.

* Change allowed file extensions

At present the docker-compose only passes through a very limited 
subset of allowed file extensions. This change brings this in line 
with what we accept within production at present.

* Add management-ui to docker compose

One of the missing services is management-ui from the docker 
compose file. Too allow us to quickly set this up for developer 
usages this has been added to the docker compose file.

* Change to use port 8080

Currently we are using port 8099 internally for the management-ui, 
all the other services use 8080. This changes brings this in line with 
the other services to use the same port.

* Improve environment and depends_on

A number of the environment variables that are currently passed in 
don't need to be. This is either because sensible defaults are set or 
the variables are not used. Further tickets should be created to 
identify any changing defaults required. 
Alongside this, the workflow service is not required for at least the 
initial use case of management-ui so this is removed. We do 
however require that the s3 buckets in localstack have been created 
to use the system effectively.